pyproject.toml
[project]
name = "phishing-radar"
version = "0.1.0"
description = "Real-time phishing infrastructure detector from Certificate Transparency logs"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.6.0",
    "websockets>=13.0",
    "certifi>=2024.8.30",
    "rapidfuzz>=3.10.0",
    "google-cloud-bigquery>=3.26.0",
    "google-cloud-storage>=2.18.0",
    "dlt[bigquery,filesystem]>=1.4.0",
    "requests>=2.32.0",
    "pandas>=2.2.0",
    "pyarrow>=18.0.0",
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
    "pydeck>=0.9.1",
    "maxminddb>=2.6.2",
    "stix2>=3.0.1",
    "db-dtypes>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "dbt-bigquery>=1.11.1",
    "dbt-core>=1.11.8",
]
